VMCall Interface
This provides the support for 'vmcall'. BFM has been modified
to support the vmcall interface, but custom applications can
use this interface to do whatever they want, including calling
vmcall from user-space if so desired.

This patch also provides new code in the memory manager to
handle mapping memory from the Host OS into the hypervisor.

Finally, this patch fixes a host of bugs including a bug
with TLS, and cleans up the code that was touched in
prep for 1.1

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/182
[ISSUE]: https://github.com/Bareflank/hypervisor/issues/274

Signed-off-by: “Rian <“rianquinn@gmail.com”>