Delegate Implementation
This patch cleans up some issues with formatting, and adds support
for C++ delegates. This implementation doesn't need malloc/free
and is as fast as a regular vtable lookup. Before the release
of v1.2, a lot of the inheritance currently being used with be
converted to delegates to increase modularity.

Signed-off-by: “rianquinn” <“rianquinn@gmail.com”>