Naming suggestions
1) Removed all references to the word "ioctl" in the public interface,
because "ioctl" is by nature an implementation detail. Suggesting
"vmmctl" as a more generic name that collects together all interfaces
that exert control over a vmm

2) Removed the "*::interface" namespace in favor of exposing vmmctl
namespace to users of the library, and using vmmctl::details for
internal development

3) Suggesting we replace the top level "host::" namespace with
"vmmctl::". Then microV could also have a "vmctl" namespaces for working
with VM's

Signed-off-by: JaredWright <jared.wright12@gmail.com>